+++ example/SimplePingBot/Program.cs
@@ -10,7 +10,10 @@ string auth = config.Value<string>("auth")!,
 
 using GuildedBotClient client = new(auth);
 
-client.Prepared += (_, _) => Console.WriteLine("The bot is prepared!");
+client.Prepared
+    .Subscribe(me =>
+        Console.WriteLine("The bot is prepared!\nLogged in as \"{0}\" with an id \"{1}\"", me.Name, me.Id)
+    );
 
 // Wait for !ping messages
 client.MessageCreated