+++ src/Guilded.Base/user/Me.cs
@@ -1,70 +1,68 @@
-// namespace Guilded.Base.Users
-// {
-//     /// <summary>
-//     /// Information about this user.
-//     /// </summary>
-//     public class Me : ClientObject
-//     {
-//         #region JSON properties
-//         /// <summary>
-//         /// List of teams/guilds/server this user is currently in.
-//         /// </summary>
-//         /// <value>List of teams</value>
-//         [JsonProperty(Required = Required.Always)]
-//         public IList<BaseTeam> Teams { get; }
-//         /// <summary>
-//         /// The user itself.
-//         /// </summary>
-//         /// <value>User</value>
-//         [JsonProperty(Required = Required.Always)]
-//         public ThisUser User { get; }
-//         /// <summary>
-//         /// Custom emotes which can be used in Guilded by this user.
-//         /// </summary>
-//         /// <value>List of emotes</value>
-//         [JsonProperty("customReactions", Required = Required.Always)]
-//         public IList<BaseEmote> CustomEmotes { get; }
-//         /// <summary>
-//         /// How many times these emotes have been used.
-//         /// </summary>
-//         /// <value></value>
-//         [JsonProperty("reactionUsages", Required = Required.Always)]
-//         public IList<EmoteUse> EmoteUses { get; }
-//         /// <summary>
-//         /// A list of friends of this user, friend requests sent by user and friend requests sent to user.
-//         /// </summary>
-//         /// <value>List of friends</value>
-//         [JsonProperty(Required = Required.Always)]
-//         public IList<Friend> Friends { get; }
-//         #endregion
+using System;
+using Newtonsoft.Json;
 
-//         #region Additional
-//         /// <summary>
-//         /// Username of this user.
-//         /// </summary>
-//         /// <value>Name</value>
-//         [JsonIgnore]
-//         public string Username => User.Username;
-//         /// <summary>
-//         /// ID of this user.
-//         /// </summary>
-//         /// <value>User ID</value>
-//         [JsonIgnore]
-//         public Hashid Id => User.Id;
-//         /// <summary>
-//         /// Count of how many teams this client is in.
-//         /// </summary>
-//         /// <value>Count</value>
-//         [JsonIgnore]
-//         public int TeamCount => Teams.Count;
+namespace Guilded.Base.Users;
 
-//         /// <summary>
-//         /// Checks if this user is in a specific team.
-//         /// </summary>
-//         /// <param name="teamId">ID of the team</param>
-//         /// <returns>Is in a team with given ID</returns>
-//         public bool InTeam(HashId teamId) =>
-//             Teams.Any(team => team.Id == teamId);
-//         #endregion
-//     }
-// }
+/// <summary>
+/// The information about the current logged in user.
+/// </summary>
+public class Me : BaseObject
+{
+    /// <summary>
+    /// The identifier of this user.
+    /// </summary>
+    /// <remarks>
+    /// <para>The user identifier of the current client.</para>
+    /// </remarks>
+    /// <value>User ID</value>
+    public HashId Id { get; set; }
+    /// <summary>
+    /// The identifier of this bot.
+    /// </summary>
+    /// <remarks>
+    /// <para>The bot identifier of the current client to distinguish it from flowbots.</para>
+    /// </remarks>
+    /// <value>Bot ID</value>
+    public Guid BotId { get; set; }
+    /// <summary>
+    /// The name of this client.
+    /// </summary>
+    /// <value>Name</value>
+    public string Name { get; set; }
+    /// <summary>
+    /// The creation date of this client.
+    /// </summary>
+    /// <value>Date</value>
+    public DateTime CreatedAt { get; set; }
+    /// <summary>
+    /// The identifier of the creator of this bot.
+    /// </summary>
+    /// <value>User ID</value>
+    public HashId CreatedBy { get; set; }
+    /// <summary>
+    /// Creates new instance of <see cref="Me" /> with the provided details.
+    /// </summary>
+    /// <param name="id">The identifier of this user</param>
+    /// <param name="botId">The identifier of this bot</param>
+    /// <param name="name">The name of this client</param>
+    /// <param name="createdAt">The creation date of this client</param>
+    /// <param name="createdBy">The identifier of the creator of this bot</param>
+    [JsonConstructor]
+    public Me(
+        [JsonProperty(Required = Required.Always)]
+        HashId id,
+
+        [JsonProperty(Required = Required.Always)]
+        Guid botId,
+
+        [JsonProperty(Required = Required.Always)]
+        string name,
+
+        [JsonProperty(Required = Required.Always)]
+        DateTime createdAt,
+
+        [JsonProperty(Required = Required.Always)]
+        HashId createdBy
+    ) =>
+        (Id, BotId, Name, CreatedAt, CreatedBy) = (id, botId, name, createdAt, createdBy);
+}